requirements.txt
Django>=1.11.3,<2.0
psycopg2==2.7.4
django-taggit==0.22.2
django-taggit-autocomplete-modified==0.1.0b6
django_compressor==2.2
django-mptt==0.8.7
social-auth-app-django==2.1.0
Pillow==5.1.0
psycopg2==2.7.4
djangorestframework==3.8.2
django-extra-fields==1.0.0
